添加使用Dockerfile 构建依赖库，容器镜像包含swoole-cli运行时 (#106)
* 添加使用Dockerfile 构建依赖库

* 移除无用文件

* 移除无用文件

* 更新swoole-cli 版本

* 使用最新的依赖库

* disable libidn2

* 改进 生成 容器镜像 docker.io/phpswoole/swoole-cli-builder

+++ sapi/download-box/download-box-get-archive-from-server.sh
@@ -23,9 +23,9 @@ DOMAIN='http://127.0.0.1:8000'
 DOMAIN='https://swoole-cli.jingjingxyk.com/'
 URL="${DOMAIN}/all-archive.zip"
 
-wget -O all-archive.zip ${URL}
+test -f  all-archive.zip || wget -O all-archive.zip ${URL}
 
-unzip all-archive.zip
+unzip -n all-archive.zip
 
 cd ${__PROJECT__}/
